+++ core/src/main/java/org/example/reporters/HttpMetricsReporter.java
@@ -12,7 +12,7 @@ public class HttpMetricsReporter implements MetricsReporter<HttpRequestTimer> {
     }
 
     @Override
-    public void report(HttpRequestTimer annotation, String methodName, double durationNanos) {
+    public void report(HttpRequestTimer annotation, String methodName, long durationNanos) {
         registry.record("HTTP Route", annotation.path(), durationNanos);
     }
 }
